Move < 10.4 OS X rejection code to App. Don't burden UnicornApplication.
git-svn-id: svn+ssh://svn.last.fm/svn/clientside/trunk/desktop@119929 ab8f4a45-97f9-0310-bbd1-854ce3dcee89

+++ app/clientplugins/localresolver/LocalCollection.cpp
@@ -783,7 +783,7 @@ LocalCollection::getFilesToTag(int maxTagAgeDays)
         "INNER JOIN artists ON artists.id = files.artist "
         "WHERE tag_time IS NULL " 
         "OR tag_time < :oldTagAge").
-        bindValue(":oldTagAge", oldTagAge).
+        bindValue(":oldTagAge", (uint)oldTagAge).
         setForwardOnly( true ).
         exec();
     QList<FilesToTagResult> results;
